+++ packages/autocomplete-client/src/package-version.ts
@@ -1,1 +1,1 @@
-export const PACKAGE_VERSION = '0.2.2'
+export const PACKAGE_VERSION = '0.3.0'