+++ packages/instant-meilisearch/src/types/types.ts
@@ -54,6 +54,7 @@ export type OverridableMeiliSearchSearchParameters = Pick<
   | 'matchingStrategy'
   | 'showRankingScore'
   | 'attributesToSearchOn'
+  | 'hybrid'
 >
 
 type BaseInstantMeiliSearchOptions = {