+++ packages/instant-meilisearch/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @meilisearch/instant-meilisearch
 
+## 0.16.0
+
+### Minor Changes
+
+- b1343c1: Enable experimental hybrid search
+
+## 0.15.0
+
+### Minor Changes
+
+- 72726ac: Update meilisearch-js version
+- 36b5a62: Fixed RegExp in filter-adapter.ts and sort-context.ts to work in Safari
+
 ## 0.14.0
 
 ### Minor Changes